fix(frontend): apply box-shadow and border-radius to same elements
Having the box-shadow and border-radius on two different elements
resulted in a small gap being visible between a rectangular box shadow
and an image with rounded corners. The `sx` prop does not allow
targeting the underlying `Image` component directly (which is what the
`radius` prop is targeting), but the mantine Styles API (used via the
`styles` prop) does.